File formats tidy and extend a test class for cams writer. Additional tests
