+++ features/environment.py
@@ -15,13 +15,13 @@
 #  ====================================================================
 
 from behave import fixture
-from inat_to_cams import cams_interface, cams_reader, cams_writer, synchronise_inat_to_cams
+from inat_to_cams import cams_interface, cams_reader, test_cams_writer, synchronise_inat_to_cams
 import time
 
 
 @fixture
 def before_all(context):
-    context.writer = cams_writer.CamsWriter()
+    context.writer = test_cams_writer.TestCamsWriter()
     context.reader = cams_reader.CamsReader()
     context.synchroniser = synchronise_inat_to_cams.synchroniser
     context.connection = cams_interface.connection
